requirements.txt
tensorflow>=2.5.0*
keras==2.4.3
imutils==0.5.4
numpy==1.19.5
opencv-python>=4.2.0.32
matplotlib==3.4.1
argparse==1.4.0
scipy==1.6.2
scikit-learn==0.24.1
pillow>=8.3.2
streamlit==0.79.0
onnx==1.10.1
tf2onnx==1.9.3
pyqt5 
